Clean up some more warnings
Also, fix the SDL makefile so that it can build with music support
without needing to link against libWildMidi.a.

+++ software/firmware/linuxdoom-1.10/d_main.c
@@ -1176,7 +1176,7 @@ void D_DoomMain (void)
 	// for statistics driver
 	extern  void*	statcopy;                            
 
-	statcopy = (void*)atoi(myargv[p+1]);
+	statcopy = (void*)(uintptr_t)atoi(myargv[p+1]);
 	printf ("External statistics registered.\n");
     }
